YAY-114 refactor: transfer EmailSentModal, OAuthErrorModal to modals
