+++ src/features/auth/ui/modals/EmailSentModal/EmailSentModal.tsx
